added other states - bug in switch from work to big break
error - one useEffect is depending on another useEffect and the effect doesn't happen in time.

one effect is adding a finished pomodoro if the session is work and the timer finishes.

one effect checks for the amount of completed pomodoros to decide whether the next session is break or big break.

the check happens before the pomodoro is added so when it is time for big break, break is scheduled instead.